test: Move tests to provider module and remove duplicates
