+++ lib/plugins/aws/deploy/lib/cleanupEcrRepository.js
@@ -29,14 +29,13 @@ module.exports = {
     }
 
     // We should only keep the images that are tagged with imageNames that are defined in service
-    const imageTagsDefinedInService = Object.keys(
-      _.get(this.serverless.service.provider, 'ecr.images', {})
+    const imageTagsDefinedInService = new Set(
+      Object.keys(_.get(this.serverless.service.provider, 'ecr.images', new Set()))
     );
     const imageDigestsToRemove = imageDetails
       .filter(
         (image) =>
-          !image.imageTags ||
-          !image.imageTags.some((tag) => imageTagsDefinedInService.includes(tag))
+          !image.imageTags || !image.imageTags.some((tag) => imageTagsDefinedInService.has(tag))
       )
       .map(({ imageDigest }) => ({ imageDigest }));
