+++ lib/plugins/aws/provider.js
@@ -32,6 +32,8 @@ const constants = {
   providerName: 'aws',
 };
 
+const imageNamePattern = '^[a-z][a-z0-9-_]{1,31}$';
+
 const apiGatewayUsagePlan = {
   type: 'object',
   properties: {
@@ -774,7 +776,7 @@ class AwsProvider {
                 images: {
                   type: 'object',
                   patternProperties: {
-                    '^[a-z]{1,32}$': {
+                    [imageNamePattern]: {
                       anyOf: [
                         {
                           type: 'object',
@@ -1007,7 +1009,7 @@ class AwsProvider {
                 { $ref: '#/definitions/ecrImageUri' },
                 {
                   type: 'string',
-                  pattern: '^[a-z]{1,32}$',
+                  pattern: imageNamePattern,
                 },
               ],
             },
@@ -1758,7 +1760,8 @@ Object.defineProperties(
           await spawnExt('docker', ['--version']);
         } catch (err) {
           throw new this.serverless.classes.Error(
-            'Could not find Docker installation. Ensure Docker is installed before continuing.'
+            'Could not find Docker installation. Ensure Docker is installed before continuing.',
+            'DOCKER_COMMAND_NOT_AVAILABLE'
           );
         }
       },
@@ -1783,7 +1786,8 @@ Object.defineProperties(
           throw new this.serverless.classes.Error(
             `Encountered error during executing: docker ${dockerArgs.join(
               ' '
-            )}\nOutput of the command:\n${err.stdBuffer}`
+            )}\nOutput of the command:\n${err.stdBuffer}`,
+            'DOCKER_LOGIN_ERROR'
           );
         }
       },
@@ -1824,18 +1828,19 @@ Object.defineProperties(
         const pathToDockerfile = path.join(imagePath, 'Dockerfile');
 
         try {
-          const stat = await fs.promises.stat(pathToDockerfile);
-          isDockerfileAvailable = stat.isFile();
+          const stats = await fs.promises.stat(pathToDockerfile);
+          isDockerfileAvailable = stats.isFile();
         } catch {
           // pass and handle after catch block
         }
         if (!isDockerfileAvailable) {
           throw new this.serverless.classes.Error(
-            `Could not access Dockerfile under path: "${pathToDockerfile}"`
+            `Could not access Dockerfile under path: "${pathToDockerfile}"`,
+            'DOCKERFILE_NOT_AVAILABLE_ERROR'
           );
         }
 
-        const [{ repositoryUri, repositoryName }, ,] = await Promise.all([
+        const [{ repositoryUri, repositoryName }] = await Promise.all([
           this.getOrCreateEcrRepository(),
           this.dockerLoginToEcr(),
         ]);
@@ -1851,7 +1856,8 @@ Object.defineProperties(
           throw new this.serverless.classes.Error(
             `Encountered error during executing: docker ${buildDockerArgs.join(
               ' '
-            )}\nOutput of the command:\n${err.stdBuffer}`
+            )}\nOutput of the command:\n${err.stdBuffer}`,
+            'DOCKER_BUILD_ERROR'
           );
         }
 
@@ -1863,7 +1869,8 @@ Object.defineProperties(
           throw new this.serverless.classes.Error(
             `Encountered error during executing: docker ${tagDockerArgs.join(
               ' '
-            )}\nOutput of the command:\n${err.stdBuffer}`
+            )}\nOutput of the command:\n${err.stdBuffer}`,
+            'DOCKER_TAG_ERROR'
           );
         }
 
@@ -1874,7 +1881,8 @@ Object.defineProperties(
           process.stdout.write(pushDockerStdBuffer);
         } catch (err) {
           throw new this.serverless.classes.Error(
-            `Encountered error during executing: docker ${pushDockerArgs}\nOutput of the command:\n${err.stdBuffer}`
+            `Encountered error during executing: docker ${pushDockerArgs}\nOutput of the command:\n${err.stdBuffer}`,
+            'DOCKER_PUSH_ERROR'
           );
         }
 
@@ -1928,12 +1936,14 @@ Object.defineProperties(
           if (_.isObject(imageDefinedInProvider)) {
             if (!imageDefinedInProvider.uri && !imageDefinedInProvider.path) {
               throw new this.serverless.classes.Error(
-                `Either "uri" or "path" property needs to be set on image "${image}"`
+                `Either "uri" or "path" property needs to be set on image "${image}"`,
+                'ECR_IMAGE_NEITHER_URI_NOR_PATH_DEFINED_ERROR'
               );
             }
             if (imageDefinedInProvider.uri && imageDefinedInProvider.path) {
               throw new this.serverless.classes.Error(
-                `Either "uri" or "path" property (not both) needs to be set on image "${image}"`
+                `Either "uri" or "path" property (not both) needs to be set on image "${image}"`,
+                'ECR_IMAGE_BOTH_URI_AND_PATH_DEFINED_ERROR'
               );
             }
             if (imageDefinedInProvider.path) {
